+++ docs/changelog.md
@@ -30,6 +30,8 @@ See [upgrade notes][1] for helpful information when upgrading from previous vers
 
 ## Unreleased
 
+## v1.44.0-B0011 (pre-release)
+
 What's changed since v1.43.0:
 
 - New features: